+++ CHANGELOG.md
@@ -5,12 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+Fixed
+- Additional fix for refresh token and logout on 401 or 403 erresponse code. [#508](https://github.com/rokwire/safer-illinois-app/issues/508)
 
 ## [2.9.8] - 2021-2-1
+### Fixed
+- Explore Shibboleth login failure due to deleted UUID [#508](https://github.com/rokwire/safer-illinois-app/issues/508)
 - Updated details of background permissions [#506](https://github.com/rokwire/safer-illinois-app/issues/506)
 - Updated details of background permissions Explore Shibboleth login failure due to deleted UUID [#508](https://github.com/rokwire/safer-illinois-app/issues/508)
 
 ## [2.9.7] - 2021-1-26
+### Fixed
 - Updated order of panel shown during onboarding [#503](https://github.com/rokwire/safer-illinois-app/issues/503)
 - Apply nextStepDate in local time when making the calculations for displayNextStepDate.